Reorganize notifications directory && fix some leftover props from earlier implementation
