+++ lib/ui/src/components/notifications/NotificationItem.js
@@ -43,7 +43,6 @@ export default function NotificationItem({ notification: { content, link } }) {
 
 NotificationItem.propTypes = {
   notification: PropTypes.shape({
-    icon: PropTypes.string,
     content: PropTypes.string.isRequired,
     link: PropTypes.string,
   }).isRequired,
