+++ lib/ui/src/components/sidebar/Sidebar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { styled } from '@storybook/theming';
 
-import NotificationList, { NotificationListSpacer } from './NotificationList';
+import NotificationList, { NotificationListSpacer } from '../notifications/NotificationList';
 import SidebarHeading from './SidebarHeading';
 import SidebarStories from './SidebarStories';
