+++ lib/ui/src/components/sidebar/Sidebar.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Sidebar from './Sidebar';
 import * as SidebarHeadingStories from './SidebarHeading.stories';
-import * as NotificationsListStories from './NotificationList.stories';
+import * as NotificationsListStories from '../notifications/NotificationList.stories';
 import * as SidebarStoriesStories from './SidebarStories.stories';
 
 export default {
